Extract library entrypoint and command handlers
